LPD-75910 Move FDSAdminPortletKeys to the frontend-data-set-api
The portlet key was internal to frontend-data-set-admin-web, so it could not be shared with other modules. Relocating it to the exported constants package lets the frontend-data-set-impl renderer reference the canonical constant instead of duplicating the portlet id literal, which the startup view work in LPD-75910 needs.
