+++ modules/apps/frontend-data-set/frontend-data-set-admin-web/src/main/java/com/liferay/frontend/data/set/admin/web/internal/portlet/FDSAdminPortlet.java
@@ -7,9 +7,9 @@ package com.liferay.frontend.data.set.admin.web.internal.portlet;
 
 import com.liferay.client.extension.type.manager.CETManager;
 import com.liferay.frontend.data.set.SystemFDSEntryRegistry;
-import com.liferay.frontend.data.set.admin.web.internal.constants.FDSAdminPortletKeys;
 import com.liferay.frontend.data.set.admin.web.internal.constants.FDSAdminWebKeys;
 import com.liferay.frontend.data.set.admin.web.internal.display.context.FDSAdminDisplayContext;
+import com.liferay.frontend.data.set.constants.FDSAdminPortletKeys;
 import com.liferay.frontend.data.set.serializer.FDSSerializer;
 import com.liferay.frontend.data.set.url.FDSAPIURLResolverRegistry;
 import com.liferay.object.service.ObjectDefinitionLocalService;
