+++ modules/apps/frontend-data-set/frontend-data-set-api/src/main/resources/com/liferay/frontend/data/set/constants/packageinfo
@@ -1,1 +1,1 @@
-version 2.7.0
+version 2.8.0
